bugs fixed

+++ EvoDevoFull.py
@@ -747,18 +747,20 @@ def makeOffspring(indivFit,allGenomes,allIDs,gen):
 		runDevo(newGenomes[i],newIDs[i],gen)
 
 genNumb = makeGen()
-dataPath = "Generation"+str(genNumb-1)+"/Data/"
-numbData = len(os.listdir(dataPath))
-if numbData > 1:
-	print "Successfully Created Generation " + str(genNumb)
+if genNumb != 1:
+	dataPath = "Generation"+str(genNumb-1)+"/Data/"
+	numbData = len(os.listdir(dataPath))
+	if numbData > 1:
+		print "Successfully Created Generation " + str(genNumb)
+	else:
+		print "Please Add Datalogs to " + dataPath
+		os.rmdir("Generation"+str(genNumb)+"/Data/")
+		os.rmdir("Generation"+str(genNumb)+"/Params/")
+		os.remove("Generation"+str(genNumb)+"/Genomes/genomes.npy")
+		os.remove("Generation"+str(genNumb)+"/Genomes/genomes.txt")
+		os.rmdir("Generation"+str(genNumb)+"/Genomes/")
+		os.rmdir("Generation"+str(genNumb))
 else:
-	print "Please Add Datalogs to " + dataPath
-	os.rmdir("Generation"+str(genNumb)+"/Data/")
-	os.rmdir("Generation"+str(genNumb)+"/Params/")
-	os.remove("Generation"+str(genNumb)+"/Genomes/genomes.npy")
-	os.remove("Generation"+str(genNumb)+"/Genomes/genomes.txt")
-	os.rmdir("Generation"+str(genNumb)+"/Genomes/")
-	os.rmdir("Generation"+str(genNumb))
-
+	print "Successfully Created Generation " + str(genNumb)
